Backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
google-generativeai
pydantic
python-dotenv 